horizon foncitonnel pout topt mais pas kopt

+++ src/frameworks/faulttolerance/dcop/algo/topt/AlgoKOptAPO.java
@@ -148,7 +148,7 @@ public class AlgoKOptAPO extends LockingBasicAlgorithm {
 						v.fixed = false;
 						for (double[] enc : lmsg.data) {
 							if (enc[0] == v.id) {
-								Variable n = view.varMap.get(enc[2]);
+								Variable n = view.varMap.get((int)enc[2]);
 								if (n == null) {
 									n = new Variable((int)enc[2], (int)enc[3], view);
 									if (lmsg.ttl <= 1)
@@ -165,7 +165,7 @@ public class AlgoKOptAPO extends LockingBasicAlgorithm {
 									c.cache();
 								}
 							} else {
-								Variable n = view.varMap.get(enc[0]);
+								Variable n = view.varMap.get((int)enc[0]);
 								if (n == null) {
 									n = new Variable((int)enc[0], (int)enc[1], view);
 									if (lmsg.ttl <= 1)
@@ -620,7 +620,6 @@ public class AlgoKOptAPO extends LockingBasicAlgorithm {
 				if (dis > d)
 					d = dis;
 			}
-
 			for (Integer i : pSet) {
 				int tmp = minDis.get(i * maxId + self.id);
 				for (Integer j : pSet) {
